Remove useSubmenuTrigger hook
